+++ cwt/cose_key_interface.py
@@ -2,7 +2,6 @@ from typing import Any, Dict, List, Optional, Union
 
 from .cbor_processor import CBORProcessor
 from .const import (
-    COSE_ALGORITHMS_SYMMETRIC,
     COSE_KEY_OPERATION_VALUES,
     COSE_KEY_TYPES,
     COSE_NAMED_ALGORITHMS_SUPPORTED,
@@ -268,23 +267,3 @@ class COSEKeyInterface(CBORProcessor):
             EncodeError: Failed to derive key.
         """
         raise NotImplementedError
-
-    def _validate_context(self, context: List[Any]):
-        if len(context) != 4 and len(context) != 5:
-            raise ValueError("Invalid context information.")
-        # AlgorithmID
-        if not isinstance(context[0], int):
-            raise ValueError("AlgorithmID should be int.")
-        if context[0] not in COSE_ALGORITHMS_SYMMETRIC.values():
-            raise ValueError(f"Unsupported or unknown algorithm: {context[0]}.")
-        # PartyVInfo
-        if not isinstance(context[1], list) or len(context[1]) != 3:
-            raise ValueError("PartyUInfo should be list(size=3).")
-        # PartyUInfo
-        if not isinstance(context[2], list) or len(context[2]) != 3:
-            raise ValueError("PartyVInfo should be list(size=3).")
-        # SuppPubInfo
-        if not isinstance(context[3], list) or (
-            len(context[3]) != 2 and len(context[3]) != 3
-        ):
-            raise ValueError("SuppPubInfo should be list(size=2 or 3).")